# Conflicts:
#	.idea/workspace.xml
#	eventhubs-service/pom.xml
#	kafka-service/doc/DOCKER.md
#	kafka-service/pom.xml
#	kafka-service/runContainer.sh
#	kafka-service/runSSLContainer.sh
#	kafka-service/scripts/startKafkaDispatcher.sh
#	kafka-service/scripts/startKafkaDispatcherSSH.sh
#	kafka-service/src/main/java/edu/cornell/eipm/messaging/microservices/kafka/dispatcher/About.java
#	kafka-service/src/main/resources/logback.xml

+++ kafka-service/scripts/startKafkaDispatcherSSH.sh
@@ -2,21 +2,42 @@
 
 app_folder=$1
 
+# check if the app folder exists
+if [[ -z "$app_folder" ]]; then
+      echo "ERROR: config folder not set"
+      echo "Usage: startKafkaDispatcherSSH.sh <config folder>"
+      exit 1
+else
+      echo "Using ${app_folder} as config folder"
+fi
+
+# check if the config file is available
+if [[ ! -e ${app_folder}/application.yml ]]; then
+    echo "ERROR: file ${app_folder}/application.yml not found."
+    exit 1
+fi
+
+# check if another instance is running
 if [[ -e .pid ]]; then
     echo "A Kafka-Dispatcher instance seems already running from this folder."
     echo "Stop the container first or remove the .pid file if the container is not running."
     exit 1
 fi
-mkdir -p ${app_folder}/log/
+
+mkdir -p ${app_folder}/log/ || true
+chmod a+w+r ${app_folder}/log/
 name="kd$RANDOM"
+
 echo "$name" > .pid
 
-nohup docker run -p 9098:8080 --rm --name ${name} \
+docker run -d --rm --name ${name} \
     -e HOST_HOSTNAME=$(hostname) \
     -e HOST_USER=$LOGNAME \
     -e HOST_USER_ID=$(id -u) \
+    -e DISPATCHER_PORT=8080 \
     -v ${app_folder}/log/:/log/ \
     -v $HOME/.ssh/:/ssh/:ro \
+    --net=host \
     --userns=host \
     -v ${app_folder}/application.yml:/config/application.yml \
-    eipm/kafka-dispatcher:1.2.1 &
+    cgen/kafka-dispatcher:1.3.0
